v2.0.0.44 TextKeyFrames duration is now a double instead of an int Bugfix: 2D elements should no longer show if the CKObject is inactive Cleaned up some code regarding HitsplatOverlay generation

+++ src/main/java/com/creatorskit/swing/timesheet/keyframe/TextKeyFrame.java
@@ -7,10 +7,10 @@ import lombok.Setter;
 @Setter
 public class TextKeyFrame extends KeyFrame
 {
-    private int duration;
+    private double duration;
     private String text;
 
-    public TextKeyFrame(double tick, int duration, String text)
+    public TextKeyFrame(double tick, double duration, String text)
     {
         super(KeyFrameType.TEXT, tick);
         this.duration = duration;
